requirements.txt
Django==3.2.6
celery==5.1.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
django-rest-authtoken==2.1.4
